backend/requirements.txt
flask
flask-smorest
sqlalchemy
flask-sqlalchemy
flask-migrate
flask-jwt-extended
python-dotenv
passlib
marshmallow
requests
email-validator
psycopg2
gunicorn